actualizado de repos no asincrónicos

+++ src/Spotify.Core/Persistencia/IRepoGenero.cs
@@ -1,4 +1,4 @@
 namespace Spotify.Core.Persistencia;
 
-public interface IRepoGenero : IAlta<Genero, byte>, IListado<Genero>, IEliminar<uint>, IDetallePorId<Genero,byte>
+public interface IRepoGenero : IAlta<Genero, byte>, IListado<Genero>, IEliminar<byte>, IDetallePorId<Genero,byte>
 {}
